rename dom to dom-utils
