+++ src/dom-utils/__tests__/prefixClass.test.js
