+++ src/rules/Footnote.js
@@ -1,6 +1,6 @@
 import Replace from './Replace';
 import { validate, T } from '../option-checker';
-import { createEl } from '../dom';
+import { createEl } from '../dom-utils';
 
 // Options:
 // selector: String
