+++ src/rules/FullBleedSpread.js
@@ -1,6 +1,6 @@
 import OutOfFlow from './OutOfFlow';
 import { validate, T } from '../option-checker';
-import { c, createEl } from '../dom';
+import { c, createEl } from '../dom-utils';
 
 // Options:
 // selector: String
