+++ src/viewer/Layouts/printLayout.js
@@ -1,4 +1,4 @@
-import { createEl } from '../../dom';
+import { createEl } from '../../dom-utils';
 import { printMarksSingle, printMarksSpread, bookletMeta } from './printMarks';
 
 const twoPageSpread = children => createEl('.spread-wrapper.spread-size', children);
