+++ src/viewer/Viewer.js
@@ -1,6 +1,6 @@
 import { Page, orderPagesBooklet } from '../book';
 import { Mode, Paper, Layout, Marks } from '../main/Constants';
-import { c, createEl } from '../dom';
+import { c, createEl } from '../dom-utils';
 import { oncePerFrameLimiter } from '../utils';
 
 import errorView from './error';
