+++ src/viewer/__tests__/error.test.js
@@ -1,5 +1,5 @@
 import error from '../error';
-import { c } from '../../dom';
+import { c } from '../../dom-utils';
 
 global.BINDERY_VERSION = 'v0';
 const el = error('Title', 'Desc');
