+++ src/viewer/error.js
@@ -1,6 +1,6 @@
 /* global BINDERY_VERSION */
 
-import { createEl } from '../dom';
+import { createEl } from '../dom-utils';
 
 export default function (title, text) {
   return createEl('.error', [
